+++ mail/mail-tui/src/app_model/mailbox/composer.rs
@@ -154,9 +154,9 @@ impl Composer {
 
     fn update_draft_from_state(&mut self) -> Result<(), recipients::RecipientError> {
         // We are TUI, what else can we do?
-        self.draft.mime_type = MimeType::TextPlain;
+        self.draft.decrypted_body.metadata.mime_type = MimeType::TextPlain;
         self.draft.subject = self.subject_input_state.value().to_owned();
-        self.draft.body = self.text_area.lines().join("\n");
+        self.draft.decrypted_body.body = self.text_area.lines().join("\n");
         self.draft.cc_list = recipients_value_to_list(self.cc_input_state.value())?;
         self.draft.bcc_list = recipients_value_to_list(self.bcc_input_state.value())?;
         self.draft.to_list = recipients_value_to_list(self.to_input_state.value())?;
@@ -208,20 +208,29 @@ impl Composer {
         let to_list = recipient_list_to_display_value(&draft.to_list);
         let cc_list = recipient_list_to_display_value(&draft.cc_list);
         let bcc_list = recipient_list_to_display_value(&draft.bcc_list);
-        let text_area = if draft.mime_type == MimeType::TextHtml {
+        let text_area = if draft.decrypted_body.metadata.mime_type == MimeType::TextHtml {
             let config = html2text::config::plain();
-            let cursor = Cursor::new(&draft.body);
+            let cursor = Cursor::new(&draft.decrypted_body.body);
             let text = config
                 .string_from_read(cursor, 80)
                 .unwrap_or_else(|e| format!("Failed to parse html:{e}"));
             TextArea::new(text.split('\n').map(str::to_owned).collect())
-        } else if draft.mime_type == MimeType::TextPlain {
-            TextArea::new(draft.body.split('\n').map(str::to_owned).collect())
+        } else if draft.decrypted_body.metadata.mime_type == MimeType::TextPlain {
+            TextArea::new(
+                draft
+                    .decrypted_body
+                    .body
+                    .split('\n')
+                    .map(str::to_owned)
+                    .collect(),
+            )
         } else {
             TextArea::new(vec!["Unknown mime type".to_owned()])
         };
         let subject = draft.subject.clone();
         let attachment_infos = draft
+            .decrypted_body
+            .metadata
             .attachments
             .iter()
             .map(|attachment| AttachmentInfo {